requirements.txt
asgiref==3.9.1
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-allauth==65.10.0
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
urllib3==2.5.0
webencodings==0.5.1
